+++ src/Form/Block/ShareButtonsDisplayType.php
@@ -1,0 +1,28 @@
+<?php
+
+/*
+ * (c) 2026: 975L <contact@975l.com>
+ * (c) 2026: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace c975L\SocialBundle\Form\Block;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+// Data form of the "share_buttons_display" block kind - no fields: this block only points at the
+// site-wide "share_buttons_settings" singleton (see ShareButtonsSettingsCrudController), it holds no
+// data of its own
+class ShareButtonsDisplayType extends AbstractType
+{
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => null,
+            'translation_domain' => 'social',
+        ]);
+    }
+}